requirements.txt
streamlit
numpy
pillow
matplotlib
scikit-learn
pandas
